requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
prophet==1.1.5
plotly==5.19.0
scikit-learn==1.4.0
xgboost==2.0.3
tensorflow==2.15.0
keras==2.15.0
python-dotenv==1.0.1
requests==2.31.0
openpyxl==3.1.2
sqlalchemy==2.0.27
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-dateutil==2.8.2
holidays==0.45
joblib==1.3.2
shap==0.44.0
watchdog==3.0.0
pyarrow==15.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
watchdog>=3.0.0
openpyxl>=3.1.2  # לתמיכה בקובצי Excel
sqlalchemy>=2.0.0  # לעבודה עם SQLite
plotly
lightgbm==4.6.0 